+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+ShareActivity.swift
@@ -106,17 +106,16 @@ extension BrowserViewController {
           callback: { [weak self] in
             guard let self = self, let tab = tab else { return }
 
-            if let scriptHandler = tab.getContentScript(
-              name: BraveTranslateScriptHandler.scriptName
-            )
+            let scriptHandler =
+              tab.getContentScript(name: BraveTranslateScriptHandler.scriptName)
               as? BraveTranslateScriptHandler
-            {
-              scriptHandler.presentUI(on: self)
+            if let tabHelper = scriptHandler?.tabHelper {
+              tabHelper.presentUI(on: self)
 
               if tab.translationState == .active {
-                scriptHandler.revertTranslation()
+                tabHelper.revertTranslation()
               } else if tab.translationState != .active {
-                scriptHandler.startTranslation(canShowToast: true)
+                tabHelper.startTranslation(canShowToast: true)
               }
             }
           }
